Module Tezos_raw_protocol_017_PtNairob

module Alpha_context : sig ... end

module Alpha_services : sig ... end

module Amendment : sig ... end

Amendments and proposals.

module Apply : sig ... end

This module supports advancing the ledger state by applying operations.

module Apply_internal_results : sig ... end

module Apply_operation_result : sig ... end

module Apply_results : sig ... end

module Baking : sig ... end

module Bitset : sig ... end

module Blinded_public_key_hash : sig ... end

module Block_header_repr : sig ... end

module Block_payload_hash : sig ... end

module Block_payload_repr : sig ... end

Value on which validators try to reach a consensus.

module Bond_id_repr : sig ... end

module Bootstrap_storage : sig ... end

This module provides functions that can be used in a private network to delay initial rewarding, typically when waiting for more bakers to join the network.

module Bounded_history_repr : sig ... end

module Cache_memory_helpers : sig ... end

module Cache_repr : sig ... end

Frequently used data should be kept in memory and persisted along a chain of blocks. The caching mechanism allows the economic protocol to declare such data and to rely on a Least Recently Used strategy to keep the cache size under a fixed limit.

module Carbonated_map : sig ... end

An in-memory data-structure for a key-value map where all operations account for gas costs.

module Carbonated_map_costs : sig ... end

module Commitment_repr : sig ... end

module Commitment_storage : sig ... end

module Constants_parametric_previous_repr : sig ... end

module Constants_parametric_repr : sig ... end

module Constants_repr : sig ... end

module Constants_services : sig ... end

module Constants_storage : sig ... end

This module provides functions to extract the value of protocol parameters from the context. See Constant_repr.parametric for more details about these values.

module Context_binary_proof : sig ... end

module Contract_delegate_storage : sig ... end

This module deals with the delegates of a contract. It is responsible for maintaining the tables Storage.Contract.Delegate and Storage.Contract.Delegated.

module Contract_hash : sig ... end

module Contract_manager_storage : sig ... end

module Contract_repr : sig ... end

module Contract_services : sig ... end

module Contract_storage : sig ... end

module Cycle_repr : sig ... end

module Dal_apply : sig ... end

This modules handles all the validation/application/finalisation of any operation related to the DAL.

module Dal_attestation_repr : sig ... end

module Dal_errors_repr : sig ... end

module Dal_operations_repr : sig ... end

module Dal_services : sig ... end

module Dal_slot_index_repr : sig ... end

module Dal_slot_repr : sig ... end

module Dal_slot_storage : sig ... end

Storage management of slots for the data-availability layer.

module Delegate_activation_storage : sig ... end

This module deals with delegates' activity. Typically, the provided functions can be used to deactivate a delegate that has not shown activity for a certain number of cycles, and to reactivate it when appropriate.

module Delegate_consensus_key : sig ... end

Management of a delegate's consensus key, the one used to sign blocks and consensus operations. It is responsible for maintaining the tables Storage.Consensus_keys, Storage.Contract.Consensus_key, and Storage.Contract.Pending_consensus_keys.

module Delegate_cycles : sig ... end

Per-cycle management of delegates.

module Delegate_missed_endorsements_storage : sig ... end

This modules deals with delegates' participation in consensus.

module Delegate_sampler : sig ... end

This module draws random values for a cycle based on the Seed_repr.seed associated that cycle. These random values are:

module Delegate_services : sig ... end

module Delegate_slashed_deposits_storage : sig ... end

This module maintains the storage related to slashing of delegates for double signing. In particular, it is responsible for maintaining the Storage.Slashed_deposits table.

module Delegate_storage : sig ... end

This module groups everything related to delegate registration. For the invariants maintained, see the submodule Contract.

module Dependent_bool : sig ... end

Dependent booleans

module Destination_repr : sig ... end

module Destination_storage : sig ... end

module Entrypoint_repr : sig ... end

module Fees_storage : sig ... end

module Fitness_repr : sig ... end

module Fixed_point_repr : sig ... end

module Frozen_deposits_storage : sig ... end

Simple abstraction from low-level storage to handle frozen deposits.

module Gas_comparable_input_size : sig ... end

module Gas_input_size : sig ... end

module Gas_limit_repr : sig ... end

module Gas_monad : sig ... end

module Global_constants_costs : sig ... end

Costs function for the global table of constants.

module Global_constants_storage : sig ... end

This module represents access to a global table of constant Micheline values. Users may register a Micheline value in the table, paying the cost of storage. Once stored, contracts source code may reference this value by its hash.

module Indexable : sig ... end

module Init_storage : sig ... end

Functions to setup storage. Used by Alpha_context.prepare.

module Lazy_storage_diff : sig ... end

module Lazy_storage_kind : sig ... end

module Legacy_script_patches : sig ... end

module Level_repr : sig ... end

module Level_storage : sig ... end

module Liquidity_baking_cpmm : sig ... end

module Liquidity_baking_lqt : sig ... end

module Liquidity_baking_migration : sig ... end

module Liquidity_baking_repr : sig ... end

module Liquidity_baking_storage : sig ... end

module Local_gas_counter : sig ... end

This module exposes an API for local gas counting. It provides a set of functions for updating a gas counter without applying it on an an Alpha_context.context.

module Main : sig ... end

module Manager_counter_repr : sig ... end

module Manager_repr : sig ... end

module Mempool_validation : sig ... end

module Merkle_list : sig ... end

module Michelson_v1_gas : sig ... end

This module provides the gas costs for typechecking Michelson scripts, parsing and unparsing Michelson values, and interpreting Michelson instructions.

module Michelson_v1_gas_costs : sig ... end

module Michelson_v1_gas_costs_generated : sig ... end

module Michelson_v1_primitives : sig ... end

module Migration_repr : sig ... end

module Misc : sig ... end

module Non_empty_string : sig ... end

module Nonce_hash : sig ... end

module Nonce_storage : sig ... end

module Operation_costs : sig ... end

module Operation_repr : sig ... end

module Origination_nonce : sig ... end

module Parameters_repr : sig ... end

module Path_encoding : sig ... end

module Period_repr : sig ... end

module Ratio_repr : sig ... end

module Raw_context : sig ... end

module Raw_context_intf : sig ... end

module Raw_level_repr : sig ... end

module Receipt_repr : sig ... end

module Round_repr : sig ... end

module Sampler : sig ... end

module Sapling_repr : sig ... end

module Sapling_services : sig ... end

module Sapling_storage : sig ... end

module Sapling_validator : sig ... end

module Saturation_repr : sig ... end

module Sc_rollup_PVM_sig : sig ... end

module Sc_rollup_arith : sig ... end

module Sc_rollup_commitment_repr : sig ... end

module Sc_rollup_commitment_storage : sig ... end

Defines storage for Smart Contract Optimistic Rollups.

module Sc_rollup_costs : sig ... end

This module contains constants and utility functions for gas metering functions used when handling SC rollups operations in context.

module Sc_rollup_data_version_sig : sig ... end

module Sc_rollup_dissection_chunk_repr : sig ... end

module Sc_rollup_errors : sig ... end

module Sc_rollup_game_repr : sig ... end

module Sc_rollup_inbox_merkelized_payload_hashes_repr : sig ... end

module Sc_rollup_inbox_message_repr : sig ... end

module Sc_rollup_inbox_repr : sig ... end

module Sc_rollup_inbox_storage : sig ... end

module Sc_rollup_management_protocol : sig ... end

This module provides a typed API for the Rollup Management Protocol that defines the communication protocol for exchanging messages between Layer 1 and Layer 2 for smart-contract rollups.

module Sc_rollup_metadata_repr : sig ... end

module Sc_rollup_operations : sig ... end

module Sc_rollup_outbox_message_repr : sig ... end

module Sc_rollup_outbox_storage : sig ... end

A module for managing state concerning a rollup's outbox.

module Sc_rollup_proof_repr : sig ... end

module Sc_rollup_refutation_storage : sig ... end

module Sc_rollup_repr : sig ... end

module Sc_rollup_reveal_hash : sig ... end

module Sc_rollup_stake_storage : sig ... end

module Sc_rollup_staker_index_repr : sig ... end

module Sc_rollup_staker_index_storage : sig ... end

module Sc_rollup_storage : sig ... end

module Sc_rollup_tick_repr : sig ... end

module Sc_rollup_wasm : sig ... end

module Sc_rollups : sig ... end

module Script_big_map : sig ... end

module Script_bytes : sig ... end

Semantics of logical and bit-shift operators for bytes

module Script_cache : sig ... end

This module manages the cache for smart contracts.

module Script_comparable : sig ... end

module Script_expr_hash : sig ... end

module Script_int : sig ... end

module Script_interpreter : sig ... end

This is the Michelson interpreter.

module Script_interpreter_defs : sig ... end

module Script_ir_annot : sig ... end

module Script_ir_translator : sig ... end

module Script_ir_translator_config : sig ... end

module Script_ir_unparser : sig ... end

module Script_list : sig ... end

module Script_map : sig ... end

Functions to ease the manipulation of Michelson maps.

module Script_repr : sig ... end

module Script_set : sig ... end

Functions to ease the manipulation of sets of values in Michelson.

module Script_string : sig ... end

Strings of printable characters

module Script_tc_context : sig ... end

This module defines the typechecking context used during the translation from Michelson untyped nodes to typed nodes (Script_ir_translator). The context keeps track of the origin of the code (top-level from a contract, in a view, etc.), plus some information to allow or forbid instructions given the context (no `SELF` in a lambda for example).

module Script_tc_errors : sig ... end

module Script_tc_errors_registration : sig ... end

module Script_timestamp : sig ... end

module Script_typed_ir : sig ... end

module Script_typed_ir_size : sig ... end

This module provides overapproximation of memory footprint for Michelson-related values.

module Script_typed_ir_size_costs : sig ... end

module Seed_repr : sig ... end

module Seed_storage : sig ... end

This modules handles the storage of random nonce seeds.

module Services_registration : sig ... end

module Skip_list_costs : sig ... end

module Skip_list_repr : sig ... end

module Slot_repr : sig ... end

module Stake_storage : sig ... end

This module provides basic operations (accessors and setters) on staking tokens.

module State_hash : sig ... end

module Storage : sig ... end

module Storage_costs : sig ... end

module Storage_description : sig ... end

module Storage_functors : sig ... end

module Storage_sigs : sig ... end

module Tez_repr : sig ... end

module Ticket_accounting : sig ... end

module Ticket_amount : sig ... end

module Ticket_balance_key : sig ... end

This module exposes a function for generating a ticket-balance key-hash given an owner and a ticket-token. The key-hash is used for populating the global ticket-balance table that tracks ownership of tickets for different tokens.

module Ticket_costs : sig ... end

This module contains constants and utility functions for gas metering functions used for extracting and handling tickets for the global ticket balance table.

module Ticket_hash_builder : sig ... end

module Ticket_hash_repr : sig ... end

module Ticket_lazy_storage_diff : sig ... end

module Ticket_operations_diff : sig ... end

A module that provides functionality for extracting ticket-token differences from a list of operations.

module Ticket_receipt : sig ... end

module Ticket_scanner : sig ... end

This module provides an API for extracting tickets of arbitrary types from an OCaml values, given a type-witness.

module Ticket_storage : sig ... end

module Ticket_token : sig ... end

module Ticket_token_map : sig ... end

A module exposing a carbonated map where keys are Ticket_token.ex_token values.

module Ticket_token_unparser : sig ... end

module Ticket_transfer : sig ... end

module Time_repr : sig ... end

module Token : sig ... end

The aim of this module is to manage operations involving tokens such as minting, transferring, and burning. Every constructor of the types giver, container, or receiver represents a kind of account that holds a given (or possibly infinite) amount of tokens.

module Tx_rollup_commitment_repr : sig ... end

module Tx_rollup_commitment_storage : sig ... end

This module introduces various functions to manipulate the storage related to commitments for transaction rollups.

module Tx_rollup_errors_repr : sig ... end

module Tx_rollup_gas : sig ... end

module Tx_rollup_hash_builder : sig ... end

module Tx_rollup_inbox_repr : sig ... end

module Tx_rollup_inbox_storage : sig ... end

Functions to manipulate transaction rollup’s inboxes.

module Tx_rollup_l2_address : sig ... end

module Tx_rollup_l2_apply : sig ... end

module Tx_rollup_l2_batch : sig ... end

module Tx_rollup_l2_context : sig ... end

module Tx_rollup_l2_context_hash : sig ... end

module Tx_rollup_l2_context_sig : sig ... end

module Tx_rollup_l2_proof : sig ... end

module Tx_rollup_l2_qty : sig ... end

module Tx_rollup_l2_storage_sig : sig ... end

module Tx_rollup_l2_verifier : sig ... end

module Tx_rollup_level_repr : sig ... end

module Tx_rollup_message_hash_repr : sig ... end

module Tx_rollup_message_repr : sig ... end

module Tx_rollup_message_result_hash_repr : sig ... end

module Tx_rollup_message_result_repr : sig ... end

module Tx_rollup_parameters : sig ... end

A module for representing and extracting typed transactional rollup parameters.

module Tx_rollup_prefixes : sig ... end

module Tx_rollup_repr : sig ... end

module Tx_rollup_reveal_repr : sig ... end

module Tx_rollup_reveal_storage : sig ... end

module Tx_rollup_services : sig ... end

module Tx_rollup_state_repr : sig ... end

module Tx_rollup_state_storage : sig ... end

A collection of functions to manipulate the state of a transaction rollup.

module Tx_rollup_storage : sig ... end

module Tx_rollup_withdraw_list_hash_repr : sig ... end

module Tx_rollup_withdraw_repr : sig ... end

module Validate : sig ... end

module Validate_errors : sig ... end

module Vote_repr : sig ... end

module Vote_storage : sig ... end

module Voting_period_repr : sig ... end

module Voting_period_storage : sig ... end

module Voting_services : sig ... end

module Zk_rollup_account_repr : sig ... end

module Zk_rollup_apply : sig ... end

This module handles all the validation/application of any operation related to the ZK Rollup. All of the functions defined in this module require that the ZKRU feature flag is enabled.

module Zk_rollup_circuit_public_inputs_repr : sig ... end

Abstraction layer for the public inputs to the ZKRU aPlonk circuits.

module Zk_rollup_errors : sig ... end

module Zk_rollup_operation_repr : sig ... end

module Zk_rollup_parameters : sig ... end

A module for representing and extracting typed ZK rollup parameters.

module Zk_rollup_repr : sig ... end

module Zk_rollup_scalar : sig ... end

Representation of scalars used by the ZK Rollup alongside manipulation functions

module Zk_rollup_state_repr : sig ... end

module Zk_rollup_storage : sig ... end

module Zk_rollup_ticket_repr : sig ... end

module Zk_rollup_update_repr : sig ... end