Module Tezos_raw_protocol_017_PtNairob.Apply_operation_result

type ('kind, 'manager, 'successful) operation_result =

| Applied of 'successful

| Backtracked of 'successful * Tezos_protocol_environment_017_PtNairob.Error_monad.error Tezos_protocol_environment_017_PtNairob.Error_monad.trace option

| Failed : 'manager * Tezos_protocol_environment_017_PtNairob.Error_monad.error Tezos_protocol_environment_017_PtNairob.Error_monad.trace -> ('kind, 'manager, 'successful) operation_result

| Skipped : 'manager -> ('kind, 'manager, 'successful) operation_result

The result of an operation in the queue. Skipped ones should always be at the tail, and after a single Failed. * The 'kind parameter is the operation kind (a transaction, an origination, etc.). * The 'manager parameter is the type of manager kinds. * The 'successful parameter is the type of successful operations. The 'kind parameter is used to make the type a GADT, but 'manager and 'successful are used to share operation_result between internal and external operation results, and are instantiated for each case.

val trace_encoding : Tezos_protocol_environment_017_PtNairob.Error_monad.error Tezos_protocol_environment_017_PtNairob.Error_monad.trace Tezos_protocol_environment_017_PtNairob.Data_encoding.t