Module Tezos_raw_protocol_017_PtNairob.Bounded_history_repr

A bounded cache associating values to keys.

This data structure is basically a bounded association table that stores (a finite number of) given (key, value), with the following properties:

The insertion ordering is remembered / important. When the structure is full, older entries are removed to insert new ones;

Stored keys are unique in the data-structure.

module type NAME = sig ... end

module type KEY = sig ... end

The required interface for keys stored in the table.

module type VALUE = sig ... end

The required interface for values stored in the table.

module type S = sig ... end

The exported interface of the data structure.

module Make (Name : NAME) (Key : KEY) (Value : VALUE) : S with type key = Key.t and type value = Value.t