Module Tezos_raw_protocol_017_PtNairob.Cache_memory_helpers

module type SNodes = sig ... end

module Nodes : SNodes

The Nodes module is used to count the number of computation steps performed when evaluating the size of the in-memory graph corresponding to an OCaml value.

Helpers to deal with computing the in-memory size of values

type sint = Saturation_repr.may_saturate Saturation_repr.t

type nodes_and_size = Nodes.t * sint

val (!!) : int -> Saturation_repr.may_saturate Saturation_repr.t

val (+!) : 'a Saturation_repr.t -> 'b Saturation_repr.t -> Saturation_repr.may_saturate Saturation_repr.t

val (+?) : 'a Saturation_repr.t -> int -> Saturation_repr.may_saturate Saturation_repr.t

val (*?) : 'a Saturation_repr.t -> int -> Saturation_repr.may_saturate Saturation_repr.t

val (/?) : 'a Saturation_repr.t -> int -> 'a Saturation_repr.t

val (++) : (Nodes.t * 'a Saturation_repr.t) -> (Nodes.t * 'b Saturation_repr.t) -> Nodes.t * Saturation_repr.may_saturate Saturation_repr.t

val zero : Nodes.t * Saturation_repr.may_saturate Saturation_repr.t

val word_size : Saturation_repr.may_saturate Saturation_repr.t

val header_size : Saturation_repr.may_saturate Saturation_repr.t

val int32_size : Saturation_repr.may_saturate Saturation_repr.t

val int64_size : Saturation_repr.may_saturate Saturation_repr.t

val h1w : Saturation_repr.may_saturate Saturation_repr.t

val h2w : Saturation_repr.may_saturate Saturation_repr.t

val h3w : Saturation_repr.may_saturate Saturation_repr.t

val h4w : Saturation_repr.may_saturate Saturation_repr.t

val h5w : Saturation_repr.may_saturate Saturation_repr.t

val hh3w : Saturation_repr.may_saturate Saturation_repr.t

val hh6w : Saturation_repr.may_saturate Saturation_repr.t

val hh8w : Saturation_repr.may_saturate Saturation_repr.t

val z_size : Tezos_protocol_environment_017_PtNairob.Z.t -> Saturation_repr.may_saturate Saturation_repr.t

val string_size_gen : int -> Saturation_repr.may_saturate Saturation_repr.t

val bytes_size : bytes -> Saturation_repr.may_saturate Saturation_repr.t

val string_size : string -> Saturation_repr.may_saturate Saturation_repr.t

val blake2b_hash_size : Saturation_repr.may_saturate Saturation_repr.t

val public_key_hash_in_memory_size : Saturation_repr.may_saturate Saturation_repr.t

val ret_adding : ('a * 'b Saturation_repr.t) -> 'c Saturation_repr.t -> 'a * Saturation_repr.may_saturate Saturation_repr.t

val ret_succ_adding : (Nodes.t * 'a Saturation_repr.t) -> 'b Saturation_repr.t -> Nodes.t * Saturation_repr.may_saturate Saturation_repr.t

val ret_succ : (Nodes.t * 'a) -> Nodes.t * 'a

val option_size : ('a -> 'b Saturation_repr.t) -> 'a option -> Saturation_repr.may_saturate Saturation_repr.t

val option_size_vec : ('a -> Nodes.t * 'b Saturation_repr.t) -> 'a option -> Nodes.t * Saturation_repr.may_saturate Saturation_repr.t

val list_cell_size : 'a Saturation_repr.t -> Saturation_repr.may_saturate Saturation_repr.t

val list_fold_size : ('a -> Nodes.t * 'b Saturation_repr.t) -> 'a list -> Nodes.t * Saturation_repr.may_saturate Saturation_repr.t

val boxed_tup2 : 'a Saturation_repr.t -> 'b Saturation_repr.t -> Saturation_repr.may_saturate Saturation_repr.t

val node_size : ('a, 'b) Tezos_protocol_environment_017_PtNairob.Micheline.node -> Nodes.t * Saturation_repr.may_saturate Saturation_repr.t

val expr_size : 'a Tezos_protocol_environment_017_PtNairob.Micheline.canonical -> Nodes.t * Saturation_repr.may_saturate Saturation_repr.t