Module type Carbonated_map.S

type 'a t

type key

The type of keys in the map.

type context

The type used for the context.

val empty : 'a t

empty an empty map.

val singleton : key -> 'a -> 'a t

singleton k v returns a map with a single key k and value v pair.

val size : 'a t -> int

size m returns the number of elements of the map m in constant time.

val find : context -> key -> 'a t -> ('a option * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

find ctxt k m looks up the value with key k in the given map m and also consumes the gas associated with the lookup. The complexity is logarithmic in the size of the map.

val update : context -> key -> (context -> 'a option -> ('a option * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult) -> 'a t -> ('a t * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

update ctxt k f map updates or adds the value of the key k using f. The function accounts for the gas cost for finding the element. The updating function f should also account for its own gas cost. The complexity is logarithmic in the size of the map.

val to_list : context -> 'a t -> ((key * 'a) list * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

to_list m transforms a map m into a list. It also accounts for the gas cost for traversing the elements. The complexity is linear in the size of the map.

val of_list : context -> merge_overlap: (context -> 'a -> 'a -> ('a * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult) -> (key * 'a) list -> ('a t * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

of_list ctxt ~merge_overlaps m creates a map from a list of key-value pairs. In case there are overlapping keys, their values are combined using the merge_overlap function. The function accounts for gas for traversing the elements. merge_overlap should account for its own gas cost. The complexity is n * log n in the size of the list.

val merge : context -> merge_overlap: (context -> 'a -> 'a -> ('a * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult) -> 'a t -> 'a t -> ('a t * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

merge ctxt ~merge_overlap m1 m2 merges the maps m1 and m2. In case there are overlapping keys, their values are combined using the merge_overlap function. Gas costs for traversing all elements from both maps are accounted for. merge_overlap should account for its own gas cost. The complexity is n * log n, where n is size m1 + size m2.

val map_e : context -> (context -> key -> 'a -> ('b * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult) -> 'a t -> ('b t * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

map_e ctxt f m maps over all key-value pairs in the map m using the function f. It accounts for gas costs associated with traversing the elements. The mapping function f should also account for its own gas cost. The complexity is linear in the size of the map m.

val fold_e : context -> (context -> 'state -> key -> 'value -> ('state * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult) -> 'state -> 'value t -> ('state * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

fold_e ctxt f z m folds over the key-value pairs of the given map m, accumulating values using f, with z as the initial state. The function f must account for its own gas cost. The complexity is linear in the size of the map m.

val fold_es : context -> (context -> 'state -> key -> 'value -> ('state * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t) -> 'state -> 'value t -> ('state * context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Lwt-aware variant of fold_e.