Module Tezos_raw_protocol_017_PtNairob.Destination_storage

val must_exist : Raw_context.t -> Destination_repr.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

must_exist ctxt dest checks whether the given destination dest exists in the context ctxt. If the destination exists, a new context is returned with gas consumed for the lookup cost. If it does not exist, an error is returned.