Module Tezos_raw_protocol_017_PtNairob.Fees_storage

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Cannot_pay_storage_fee

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Negative_storage_input

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Operation_quota_exceeded

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Storage_limit_too_high

val record_global_constant_storage_space : Raw_context.t -> Tezos_protocol_environment_017_PtNairob.Z.t -> Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t

record_global_constant_storage_space ctxt size records paid storage space for registering a new global constant. Cost is <size> in bytes + 65 additional bytes for the key hash of the expression. Returns new context and the cost.

val record_paid_storage_space : Raw_context.t -> Contract_hash.t -> (Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t * Tezos_protocol_environment_017_PtNairob.Z.t) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

record_paid_storage_space ctxt contract updates the amount of storage consumed by the contract. This total size is considered as accounted for as far as future payment is concerned.

Returns a new context, the total space consumed by the contract, and the additional (and unpaid) space consumed since the last call of this function on this contract.

val check_storage_limit : Raw_context.t -> storage_limit:Tezos_protocol_environment_017_PtNairob.Z.t -> unit Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

check_storage_limit ctxt ~storage_limit raises the Storage_limit_too_high error iff storage_limit is negative or greater the constant hard_storage_limit_per_operation.

val burn_storage_fees : ?origin:Receipt_repr.update_origin -> Raw_context.t -> storage_limit:Tezos_protocol_environment_017_PtNairob.Z.t -> payer:Token.giver -> Tezos_protocol_environment_017_PtNairob.Z.t -> (Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t * Receipt_repr.balance_updates) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

burn_storage_fees ctxt ~storage_limit ~payer consumed takes funds from the payer to pay the cost of the consumed storage. This function has an optional parameter ~origin that allows to set the origin of returned balance updates (by default the parameter is set to Block_application). Returns an updated context, an updated storage limit equal to storage_limit - consumed, and the relevant balance updates. Raises the Operation_quota_exceeded error if storage_limit < consumed. Raises the Cannot_pay_storage_fee error if the funds from the payer are not sufficient to pay the storage fees.

val burn_storage_increase_fees : ?origin:Receipt_repr.update_origin -> Raw_context.t -> payer:Token.giver -> Tezos_protocol_environment_017_PtNairob.Z.t -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

burn_storage_increase_fees ctxt ~payer amount_in_bytes takes funds from the payer to pay the cost of the amount_in_bytes storage. This function has an optional parameter ~origin that allows to set the origin of returned balance updates (by default the parameter is set to Block_application). Returns an updated context and the relevant balance updates. Raises the Negative_storage_input error if the amount_in_bytes is null or negative. Raises the Cannot_pay_storage_fee error if the funds from the payer are not sufficient to pay the storage fees.

val burn_origination_fees : ?origin:Receipt_repr.update_origin -> Raw_context.t -> storage_limit:Tezos_protocol_environment_017_PtNairob.Z.t -> payer:Token.giver -> (Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t * Receipt_repr.balance_updates) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Calls burn_storage_fees with the parameter consumed mapped to the constant origination_size.

val burn_tx_rollup_origination_fees : ?origin:Receipt_repr.update_origin -> Raw_context.t -> storage_limit:Tezos_protocol_environment_017_PtNairob.Z.t -> payer:Token.giver -> (Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t * Receipt_repr.balance_updates) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Calls burn_storage_fees with the parameter consumed mapped to the constant tx_rollup_origination_size.

val burn_sc_rollup_origination_fees : ?origin:Receipt_repr.update_origin -> Raw_context.t -> storage_limit:Tezos_protocol_environment_017_PtNairob.Z.t -> payer:Token.giver -> Tezos_protocol_environment_017_PtNairob.Z.t -> (Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t * Receipt_repr.balance_updates) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

burn_sc_rollup_origination_fees ~origin ctxt ~storage_limit ~payer consumed burns the storage fees for smart contract rollup creation fees.

val burn_zk_rollup_origination_fees : ?origin:Receipt_repr.update_origin -> Raw_context.t -> storage_limit:Tezos_protocol_environment_017_PtNairob.Z.t -> payer:Token.giver -> Tezos_protocol_environment_017_PtNairob.Z.t -> (Raw_context.t * Tezos_protocol_environment_017_PtNairob.Z.t * Receipt_repr.balance_updates) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

burn_zk_rollup_origination_fees ~origin ctxt ~storage_limit ~payer consumed burns the storage fees for ZK rollup origination fees.