Module Tezos_raw_protocol_017_PtNairob.Manager_repr

type manager_key =

| Hash of Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t

| Public_key of Tezos_protocol_environment_017_PtNairob.Signature.Public_key.t

The public key of the manager of a contract is reveled only after the first operation. At Origination time, the manager provides only the hash of its public key that is stored in the contract. When the public key is actually revealed, the public key instead of the hash of the key

type t = manager_key

val encoding : t Tezos_protocol_environment_017_PtNairob.Data_encoding.encoding