Module Tezos_raw_protocol_017_PtNairob.Sc_rollup_arith

This module provides a temporary toy rollup to be used as a demo.

module type S = sig ... end

This rollup is a stack machine equipped with addition.

module Protocol_implementation : S with type context = Tezos_protocol_environment_017_PtNairob.Context.t and type state = Tezos_protocol_environment_017_PtNairob.Context.tree and type proof = Tezos_protocol_environment_017_PtNairob.Context.Proof.tree Tezos_protocol_environment_017_PtNairob.Context.Proof.t

val reference_initial_state_hash : Sc_rollup_repr.State_hash.t

This is the state hash of reference that both the prover of the node and the verifier of the protocol Protocol_implementation have to agree on (if they do, it means they are using the same tree structure).

module type P = sig ... end

module Make (Context : P) : S with type context = Context.Tree.t and type state = Context.tree and type proof = Context.proof