Module Tezos_raw_protocol_017_PtNairob.Sc_rollup_inbox_storage

val get_inbox : Raw_context.t -> (Sc_rollup_inbox_repr.t * Raw_context.t) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

get_inbox context returns the current state of the inbox, if it exists.

val add_external_messages : Raw_context.t -> string list -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

add_external_messages context messages adds messages to the smart rollups internal inbox level witness.

val add_deposit : Raw_context.t -> payload:Script_repr.expr -> sender:Contract_hash.t -> source:Tezos_protocol_environment_017_PtNairob.Signature.public_key_hash -> destination:Sc_rollup_repr.Address.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

add_deposit ~payload ~sender ~source ~destination ctxt adds the internal deposit message of payload, sender, and source to the smart-contract rollups' inbox.

See add_external_messages for returned values and failures.

val init_inbox : predecessor:Tezos_protocol_environment_017_PtNairob.Block_hash.t -> Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Initialize the inbox in the storage at protocol initialization.

val add_level_info : predecessor:Tezos_protocol_environment_017_PtNairob.Block_hash.t -> Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Adds the Info_per_level in the in-memory inbox level witness. If the current level is the first level of the current protocol then also add Migration message.

val finalize_inbox_level : Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

finalize_inbox_level ctxt ends the internal representation for the block.