Module Tezos_raw_protocol_017_PtNairob.Sc_rollup_staker_index_storage

val init : Raw_context.t -> Sc_rollup_repr.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

init ctxt rollup initialize a staker index counter for rollup.

val fresh_staker_index : Raw_context.t -> Sc_rollup_repr.t -> Tezos_protocol_environment_017_PtNairob.Signature.public_key_hash -> (Raw_context.t * Sc_rollup_staker_index_repr.t) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

fresh_staker_index ctxt rollup staker creates a new index for staker and store it in Storage.Sc_rollup.Staker_index.

val find_staker_index_unsafe : Raw_context.t -> Sc_rollup_repr.t -> Tezos_protocol_environment_017_PtNairob.Signature.public_key_hash -> (Raw_context.t * Sc_rollup_staker_index_repr.t option) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

find_staker_index_unsafe ctxt rollup staker returns the index for the rollup's staker. This function *must* be called only after they have checked for the existence of the rollup, and therefore it is not necessary for it to check for the existence of the rollup again. Otherwise, use the safe function find_staker_index.

val get_staker_index_unsafe : Raw_context.t -> Sc_rollup_repr.t -> Tezos_protocol_environment_017_PtNairob.Signature.public_key_hash -> (Raw_context.t * Sc_rollup_staker_index_repr.t) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Same as find_staker_index_unsafe but fails if the value is absent.

val remove_staker : Raw_context.t -> Sc_rollup_repr.t -> Tezos_protocol_environment_017_PtNairob.Signature.public_key_hash -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

remove_staker ctxt rollup staker cleans every storage associated to staker and it's index. The staker will be no longer considered active until a new index is given to it, through a new call to fresh_staker_index.

val list_stakers_uncarbonated : Raw_context.t -> Sc_rollup_repr.t -> Tezos_protocol_environment_017_PtNairob.Signature.public_key_hash list Tezos_protocol_environment_017_PtNairob.Lwt.t

list_stakers_uncarbonated ctxt rollup lists the active stakers on rollup.

val is_active : Raw_context.t -> Sc_rollup_repr.t -> Sc_rollup_staker_index_repr.t -> (Raw_context.t * bool) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

is_active ctxt rollup staker_index returns true iff staker_index is an active staker.

val is_staker : Raw_context.t -> Sc_rollup_repr.t -> Sc_rollup_repr.Staker.t -> (Raw_context.t * bool) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

is_staker context rollup staker returns true iff staker has a deposit on the given rollup.