Module Sc_rollups.Kind

A smart contract rollup has a kind, which assigns meaning to rollup operations.

type t =

| Example_arith

| Wasm_2_0_0

The list of available rollup kinds.

This list must only be appended for backward compatibility.

val encoding : t Tezos_protocol_environment_017_PtNairob.Data_encoding.t

val equal : t -> t -> bool

val pvm_of : t -> PVM.t

pvm_of kind returns the PVM of the given kind.

val all : t list

all returns all implemented PVM.

val of_string : string -> t option

val to_string : t -> string

val pp : Tezos_protocol_environment_017_PtNairob.Format.formatter -> t -> unit