Module Tezos_raw_protocol_017_PtNairob.Storage

Tezos Protocol Implementation - Typed storage

This module hides the hierarchical (key x value) database under pre-allocated typed accessors for all persistent entities of the tezos context.

This interface enforces no invariant on the contents of the database. Its goal is to centralize all accessors in order to have a complete view over the database contents and avoid key collisions.

module type Simple_single_data_storage = sig ... end

module Block_round : Simple_single_data_storage with type value = Round_repr.t

type deposits = {

initial_amount : Tez_repr.t;

current_amount : Tez_repr.t;

}

type missed_endorsements_info = {

remaining_slots : int;

missed_levels : int;

}

module Contract : sig ... end

Storage from this submodule must only be accessed through the module `Contract`.

module Big_map : sig ... end

module Sapling : sig ... end

module Delegates : Storage_sigs.Data_set_storage with type t := Raw_context.t and type elt = Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t

Set of all registered delegates.

module Consensus_keys : Storage_sigs.Data_set_storage with type t := Raw_context.t and type elt = Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t

Set of all active consensus keys in cycle `current + preserved_cycles + 1`

type slashed_level = {

for_double_endorsing : bool;

for_double_baking : bool;

}

module Slashed_deposits : Storage_sigs.Indexed_data_storage with type t := Raw_context.t * Cycle_repr.t and type key = Raw_level_repr.t * Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t and type value = slashed_level

Set used to avoid slashing multiple times the same event

module Stake : sig ... end

module Delegate_sampler_state : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Raw_context.consensus_pk Sampler.t and type t := Raw_context.t

State of the sampler used to select delegates. Managed synchronously with Stake.Selected_distribution_for_cycle.

Votes

module Vote : sig ... end

module type FOR_CYCLE = sig ... end

Seed

module Seed_status : Simple_single_data_storage with type value = Seed_repr.seed_status

module Seed : sig ... end

Storage from this submodule must only be accessed through the module `Seed`.

Commitments

module Commitments : Storage_sigs.Indexed_data_storage with type key = Blinded_public_key_hash.t and type value = Tez_repr.t and type t := Raw_context.t

module Ramp_up : sig ... end

Ramp up rewards

module Pending_migration : sig ... end

module Liquidity_baking : sig ... end

module Global_constants : sig ... end

A map of Script_repr.expr values, indexed by their hash (Script_expr_hash.t). Values from this map can be incorporated by any contract via the primitive Michelson_v1_primitives.H_constant.

module Ticket_balance : sig ... end

This module exposes a balance table for tracking ticket ownership. The table is a mapping from keys to values where the keys consist of a hashed representation of:

Tenderbake

module Tenderbake : sig ... end

module Tx_rollup : sig ... end

module Sc_rollup : sig ... end

module Dal : sig ... end

module Zk_rollup : sig ... end

module Legacy : sig ... end