Module Storage.Stake

module Staking_balance : Storage_sigs.Indexed_data_snapshotable_storage with type key = Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t and type value = Tez_repr.t and type snapshot = int and type t := Raw_context.t

The map of all the staking balances of all delegates, including those with less than Constants_parametric_repr.t.minimal_stake. It might be large

module Active_delegates_with_minimal_stake : Storage_sigs.Indexed_data_snapshotable_storage with type key = Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t and type value = unit and type snapshot = int and type t := Raw_context.t

This is a set, encoded in a map with value unit. This should be fairly small compared to staking balance

module Last_snapshot : Storage_sigs.Single_data_storage with type value = int and type t := Raw_context.t

Counter of stake storage snapshots taken since last cycle

module Selected_distribution_for_cycle : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = (Tezos_protocol_environment_017_PtNairob.Signature.Public_key_hash.t * Tez_repr.t) list and type t := Raw_context.t

List of active stake

module Total_active_stake : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Tez_repr.t and type t := Raw_context.t

Sum of the active stakes of all the delegates with Constants_parametric_repr.t.minimal_stake