Module C.Local_context

Local_context provides functions for local access from a specific directory.

include Raw_context_intf.VIEW with type t = local_context and type tree := Raw_context.tree and type key := Raw_context.key and type value := Raw_context.value

type t = local_context

The type for context views.

Getters

val mem : t -> Raw_context.key -> bool Tezos_protocol_environment_017_PtNairob.Lwt.t

mem t k is an Lwt promise that resolves to true iff k is bound to a value in t.

val mem_tree : t -> Raw_context.key -> bool Tezos_protocol_environment_017_PtNairob.Lwt.t

mem_tree t k is like mem but for trees.

val get : t -> Raw_context.key -> Raw_context.value Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

get t k is an Lwt promise that resolves to Ok v if k is bound to the value v in t and Storage_ErrorMissing_key otherwise.

val get_tree : t -> Raw_context.key -> Raw_context.tree Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

get_tree is like get but for trees.

val find : t -> Raw_context.key -> Raw_context.value option Tezos_protocol_environment_017_PtNairob.Lwt.t

find t k is an Lwt promise that resolves to Some v if k is bound to the value v in t and None otherwise.

val find_tree : t -> Raw_context.key -> Raw_context.tree option Tezos_protocol_environment_017_PtNairob.Lwt.t

find_tree t k is like find but for trees.

val list : t -> ?offset:int -> ?length:int -> Raw_context.key -> (string * Raw_context.tree) list Tezos_protocol_environment_017_PtNairob.Lwt.t

list t key is the list of files and sub-nodes stored under k in t. The result order is not specified but is stable.

offset and length are used for pagination.

Setters

val init : t -> Raw_context.key -> Raw_context.value -> t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

init t k v is an Lwt promise that resolves to Ok c if:

k is unbound in t;

k is bound to v in c;

and c is similar to t otherwise.

It is Storage_errorExisting_key if k is already bound in t.

val init_tree : t -> Raw_context.key -> Raw_context.tree -> t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

init_tree is like init but for trees.

val update : t -> Raw_context.key -> Raw_context.value -> t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

update t k v is an Lwt promise that resolves to Ok c if:

k is bound in t;

k is bound to v in c;

and c is similar to t otherwise.

It is Storage_errorMissing_key if k is not already bound in t.

val update_tree : t -> Raw_context.key -> Raw_context.tree -> t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

update_tree is like update but for trees.

val add : t -> Raw_context.key -> Raw_context.value -> t Tezos_protocol_environment_017_PtNairob.Lwt.t

add t k v is an Lwt promise that resolves to c such that:

k is bound to v in c;

and c is similar to t otherwise.

If k was already bound in t to a value that is physically equal to v, the result of the function is a promise that resolves to t. Otherwise, the previous binding of k in t disappears.

val add_tree : t -> Raw_context.key -> Raw_context.tree -> t Tezos_protocol_environment_017_PtNairob.Lwt.t

add_tree is like add but for trees.

val remove : t -> Raw_context.key -> t Tezos_protocol_environment_017_PtNairob.Lwt.t

remove t k v is an Lwt promise that resolves to c such that:

k is unbound in c;

and c is similar to t otherwise.

val remove_existing : t -> Raw_context.key -> t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

remove_existing t k v is an Lwt promise that resolves to Ok c if:

k is bound in t to a value;

k is unbound in c;

and c is similar to t otherwise.

val remove_existing_tree : t -> Raw_context.key -> t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

remove_existing_tree t k v is an Lwt promise that reolves to Ok c if:

k is bound in t to a tree;

k is unbound in c;

and c is similar to t otherwise.

val add_or_remove : t -> Raw_context.key -> Raw_context.value option -> t Tezos_protocol_environment_017_PtNairob.Lwt.t

add_or_remove t k v is:

add t k x if v is Some x;

remove t k otherwise.

val add_or_remove_tree : t -> Raw_context.key -> Raw_context.tree option -> t Tezos_protocol_environment_017_PtNairob.Lwt.t

add_or_remove_tree t k v is:

add_tree t k x if v is Some x;

remove t k otherwise.

Folds

val fold : ?depth:Raw_context_intf.depth -> t -> Raw_context.key -> order:[ `Sorted | `Undefined ] -> init:'a -> f: (Raw_context.key -> Raw_context.tree -> 'a -> 'a Tezos_protocol_environment_017_PtNairob.Lwt.t) -> 'a Tezos_protocol_environment_017_PtNairob.Lwt.t

fold ?depth t root ~order ~init ~f recursively folds over the trees and values of t. The f callbacks are called with a key relative to root. f is never called with an empty key for values; i.e., folding over a value is a no-op.

The depth is 0-indexed. If depth is set (by default it is not), then f is only called when the conditions described by the parameter is true:

Eq d folds over nodes and values of depth exactly d.

Lt d folds over nodes and values of depth strictly less than d.

Le d folds over nodes and values of depth less than or equal to d.

Gt d folds over nodes and values of depth strictly more than d.

Ge d folds over nodes and values of depth more than or equal to d.

If order is `Sorted (the default), the elements are traversed in lexicographic order of their keys. For large nodes, it is memory-consuming, use `Undefined for a more memory efficient fold.

Hash configurations

val config : t -> Raw_context_intf.config

config t is t's hash configuration.

val length : t -> Raw_context.key -> int Tezos_protocol_environment_017_PtNairob.Lwt.t

length t key is an Lwt promise that resolves to the number of files and sub-nodes stored under k in t.

It is equivalent to list t k >|= List.length but has a constant-time complexity.

Most of the time, this function does not perform any I/O as the length is cached in the tree. It may perform one read to load the root node of the tree in case it has not been loaded already. The initial constant is the same between list and length. They both perform the same kind of I/O reads. While list usually performs a linear number of reads, length does at most one.

val consume_gas : local_context -> Gas_limit_repr.cost -> local_context Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

Internally used in Storage_functors to consume gas from within a view. May raise Block_quota_exceeded or Operation_quota_exceeded.

val absolute_key : local_context -> Raw_context.key -> Raw_context.key

Internally used in Storage_functors to retrieve the full key of a partial key relative to the local_context.