Module Tezos_raw_protocol_017_PtNairob.Ticket_operations_diff

A module that provides functionality for extracting ticket-token differences from a list of operations.

type ticket_token_diff = private {

ticket_token : Ticket_token.ex_token;

total_amount : Script_int.n Script_int.num;

destinations : (Alpha_context.Destination.t * Script_typed_ir.ticket_amount) list;

}

A type representing ticket-token balance differences. Each value consists of:

ticket_token - the type of the ticket.

total_amount - the total amount of transferred ticket-tokens.

destinations - a list of amount and contract pairs. Invariant: total_amount is the sum of the amounts in destinations.

val ticket_diffs_of_operations : Alpha_context.context -> Script_typed_ir.packed_internal_operation list -> (ticket_token_diff list * Alpha_context.context) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

ticket_diffs_of_operations ctxt ops returns a list of ticket-tokens diffs given a context, ctxt, and list of packed operations, ops. The diffs result from either a Transaction operation with parameters containing tickets, or an Origination operation with the initial storage containing tickets.