Module Tezos_raw_protocol_017_PtNairob.Tx_rollup_gas

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Tx_rollup_negative_input_size

val hash : hash_f:(bytes list -> 'b) -> Raw_context.t -> 'a Tezos_protocol_environment_017_PtNairob.Data_encoding.t -> 'a -> (Raw_context.t * 'b) Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

A generic helper to hash an input

val hash_cost : int -> Gas_limit_repr.cost Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

hash_cost size returns the cost of gas for hashing a buffer of size bytes.

Raises Tx_rollup_negative_input_size iff size < 0.

val consume_check_path_inbox_cost : Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

val consume_check_path_commitment_cost : Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

val consume_add_message_cost : Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult

consume_add_message_cost ctxt consume the gas cost of adding a message to an inbox and return the new context.

val consume_compact_commitment_cost : Raw_context.t -> int -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult