Module Tezos_raw_protocol_017_PtNairob.Tx_rollup_l2_apply

This module introduces the interpretation (off-chain) of layer2 operations read from its inboxes.

The main concern is now to interpret operations in the layer2 context with a high-throughput and process a significant number of operations “per second.”

The operations can be crafted with indexes (see Indexable). It is meant to reduce the size of operations, and therefore, increase the number of those in a message to the layer2. You will see in this file indexables which are yet unknowns (see Indexable.unknown). They can be later on modified to indexes only (see Indexable.index_only) when the potential values have been replaced by their indexes.

Therefore, we need to have the minimal number of accesses to the context. Thus, when a value is read from its index in the context, it *must* be done once.

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Counter_mismatch of {

account : Tx_rollup_l2_address.t;

expected : int64;

provided : int64;

}

| Incorrect_aggregated_signature

| Unallocated_metadata of int32

| Multiple_operations_for_signer of Tezos_protocol_environment_017_PtNairob.Bls.Public_key.t

| Invalid_transaction_encoding

| Invalid_batch_encoding

| Unexpectedly_indexed_ticket

| Missing_ticket of Alpha_context.Ticket_hash.t

| Unknown_address of Tx_rollup_l2_address.t

| Invalid_self_transfer

| Invalid_zero_transfer

| Maximum_withdraws_per_message_exceeded of {

current : int;

maximum : int;

}

type indexes = {

address_indexes : (Tx_rollup_l2_address.t * Tx_rollup_l2_address.Indexable.index) list;

ticket_indexes : (Alpha_context.Ticket_hash.t * Tx_rollup_l2_context_sig.Ticket_indexable.index) list;

}

Applying operations in the layer2 can result in creating indexes associated to both the addresses and the ticket hashes. We keep track of these creations in order to replace the values by their indexes in future operations.

module Message_result : sig ... end

type parameters = {

tx_rollup_max_withdrawals_per_batch : int;

}

The record of parameters used during the application of messages.

module Make (Context : Tx_rollup_l2_context_sig.CONTEXT) : sig ... end