Module Tezos_raw_protocol_017_PtNairob.Tx_rollup_l2_context

val metadata_encoding : Tx_rollup_l2_context_sig.metadata Tezos_protocol_environment_017_PtNairob.Data_encoding.encoding

Type-Safe Storage Access and Gas Accounting

type _ key =

| Address_metadata : Tx_rollup_l2_context_sig.address_index -> Tx_rollup_l2_context_sig.metadata key

| Address_count : int32 key

| Address_index : Tx_rollup_l2_address.t -> Tx_rollup_l2_context_sig.address_index key

| Ticket_count : int32 key

| Ticket_index : Alpha_context.Ticket_hash.t -> Tx_rollup_l2_context_sig.ticket_index key

| Ticket_ledger : Tx_rollup_l2_context_sig.ticket_index * Tx_rollup_l2_context_sig.address_index -> Tx_rollup_l2_qty.t key

A value of type 'a key identifies a value of type 'a in an underlying, untyped storage.

This GADT is used to enforce type-safety of the abstraction of the transactions rollup context. For this abstraction to work, it is necessary to ensure that the serialization of values 'a key and 'b key cannot collide. To that end, we use Data_encoding (see packed_key_encoding).

type packed_key =

| Key : 'a key -> packed_key

A monomorphic version of Key, used for serialization purposes.

val packed_key_encoding : packed_key Tezos_protocol_environment_017_PtNairob.Data_encoding.t

The encoding used to serialize keys to be used with an untyped storage.

val value_encoding : 'a. 'a key -> 'a Tezos_protocol_environment_017_PtNairob.Data_encoding.t

value_encoding key returns the encoding to be used to serialize and deserialize values associated to a key from and to the underlying storage.

Errors

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Key_cannot_be_serialized

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Value_cannot_be_serialized

type Tezos_protocol_environment_017_PtNairob.Error_monad.error +=

| Value_cannot_be_deserialized

The Context Functor

module Make (S : Tx_rollup_l2_storage_sig.STORAGE) : Tx_rollup_l2_context_sig.CONTEXT with type t = S.t and type 'a m = 'a S.m