Module type Tx_rollup_l2_context_sig.CONTEXT

This module type describes the API of the Tx_rollup context, which is used to implement the semantics of the L2 operations.

type t

The state of the Tx_rollup context.

The context provides a type-safe, functional API to interact with the state of a transaction rollup. The functions of this module, manipulating and creating values of type t are called “context operations” afterwards.

type 'a m

The monad used by the context.

Note: It is likely to be the monad of the underlying storage. In the case of the proof verifier, as it is expected to be run into the L1, the monad will also be used to perform gas accounting. This is why all the functions of this module type needs to be inside the monad m.

module Syntax : sig ... end

The necessary monadic operators the storage monad is required to provide.

val bls_verify : (Tezos_protocol_environment_017_PtNairob.Bls.Public_key.t * bytes) list -> signature -> bool m

bls_aggregate_verify allows to verify the aggregated signature of a batch.

module Address_metadata : sig ... end

The metadata associated to an address.

module Address_index : sig ... end

Mapping between Tx_rollup_l2_address.address and address_index.

module Ticket_index : sig ... end

Mapping between Ticket_hash.t and ticket_index.

module Ticket_ledger : sig ... end

The ledger of the layer 2 where are registered the amount of a given ticket a L2 account has in its possession.