Module Tezos_raw_protocol_017_PtNairob.Tx_rollup_l2_verifier

module Verifier_storage : sig ... end

module Verifier_context : sig ... end

val verify_proof : Alpha_context.t -> Tx_rollup_l2_apply.parameters -> Alpha_context.Tx_rollup_message.t -> Tx_rollup_l2_proof.t -> proof_length:int -> agreed:Alpha_context.Tx_rollup_message_result.t -> rejected:Alpha_context.Tx_rollup_message_result_hash.t -> max_proof_size:int -> Alpha_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

verify_proof ctxt message proof ~proof_length ~agreed ~rejected ~max_proof_size verifies a Merkle proof for a L2 message, starting from the state agreed. If the proof is correct, and the final Merkle hash is not equal to rejected, then verify_proof passes.

Note that if proof_length is larger than max_proof_size and the final Merkle hash is equal to rejected, the needed proof for the rejected commitment is too large, thus, verify_proof passes and the commitment is rejected.