Module Tezos_raw_protocol_017_PtNairob.Tx_rollup_services

val state : 'a Tezos_protocol_environment_017_PtNairob.RPC_context.simple -> 'a -> Alpha_context.Tx_rollup.t -> Alpha_context.Tx_rollup_state.t Tezos_protocol_environment_017_PtNairob.Error_monad.shell_tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

val inbox : 'a Tezos_protocol_environment_017_PtNairob.RPC_context.simple -> 'a -> Alpha_context.Tx_rollup.t -> Alpha_context.Tx_rollup_level.t -> Alpha_context.Tx_rollup_inbox.t option Tezos_protocol_environment_017_PtNairob.Error_monad.shell_tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Returns the inbox for a transaction rollup for a given rollup level.

Returns Not_found if the transaction rollup exists, but does not have inbox at that level. Fails if the transaction rollup does not exist.

val commitment : 'a Tezos_protocol_environment_017_PtNairob.RPC_context.simple -> 'a -> Alpha_context.Tx_rollup.t -> Alpha_context.Tx_rollup_level.t -> Alpha_context.Tx_rollup_commitment.Submitted_commitment.t option Tezos_protocol_environment_017_PtNairob.Error_monad.shell_tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

val register : unit -> unit