Module Tezos_raw_protocol_017_PtNairob.Voting_period_storage

val init : Raw_context.t -> Voting_period_repr.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Initializes the current context with voting period information.

val init_first_period : Raw_context.t -> start_position:Tezos_protocol_environment_017_PtNairob.Int32.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Sets the initial period to {voting_period = root; kind = Proposal; start_position}.

val reset : Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Increment the index by one and set the kind to Proposal.

val succ : Raw_context.t -> Raw_context.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Increment the index by one and set the kind to its successor.

val get_current : Raw_context.t -> Voting_period_repr.t Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Returns information about the current voting period.

val get_current_kind : Raw_context.t -> Voting_period_repr.kind Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Returns the current voting period kind.

val is_last_block : Raw_context.t -> bool Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Returns true if the context level is the last of current voting period.

val blocks_before_activation : Raw_context.t -> int32 option Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

blocks_before_activation ctxt returns Some b if the current voting period is the Adoption and b blocks must be waited before activation of the next protocol amendment. Returns None if the current period is not Adoption (then more than Constants_storage.blocks_per_voting_period must be waited before activation).

val get_rpc_current_info : Raw_context.t -> Voting_period_repr.info Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Returns the voting period information for the current level.

val get_rpc_succ_info : Raw_context.t -> Voting_period_repr.info Tezos_protocol_environment_017_PtNairob.Error_monad.tzresult Tezos_protocol_environment_017_PtNairob.Lwt.t

Returns the voting period information for the next level.

module Testnet_dictator : sig ... end