Module Tezos_raw_protocol_017_PtNairob.Zk_rollup_circuit_public_inputs_repr

Abstraction layer for the public inputs to the ZKRU aPlonk circuits.

As explained in the documentation, circuits in ZKRUs will be grouped into three categories: pending (public) operations, private batches and fee circuit. Each of these expects a different set of public inputs.

type pending_op_public_inputs = {

old_state : Zk_rollup_state_repr.t;

new_state : Zk_rollup_state_repr.t;

fee : Zk_rollup_scalar.t;

exit_validity : bool;

zk_rollup : Zk_rollup_repr.t;

l2_op : Zk_rollup_operation_repr.t;

}

Public inputs expected by circuits that handle single public L2 operations.

type private_batch_public_inputs = {

old_state : Zk_rollup_state_repr.t;

new_state : Zk_rollup_state_repr.t;

fees : Zk_rollup_scalar.t;

zk_rollup : Zk_rollup_repr.t;

}

Public inputs expected by circuits that handle a batch of private L2 operations.

type fee_public_inputs = {

old_state : Zk_rollup_state_repr.t;

new_state : Zk_rollup_state_repr.t;

fees : Zk_rollup_scalar.t;

}

Public inputs expected by the circuit that handles the L2 fees.

type t =

| Pending_op of pending_op_public_inputs

| Private_batch of private_batch_public_inputs

| Fee of fee_public_inputs

val to_scalar_array : t -> Zk_rollup_scalar.t array

Conversion to the type the aPlonk verifier expects.